removed redundant configuration
