+++ platform-emf/src/main/java/com/softicar/platform/emf/table/IEmfTable.java
@@ -30,8 +30,7 @@ import com.softicar.platform.emf.authorizer.EmfAuthorizer;
 import com.softicar.platform.emf.authorizer.IEmfAttributeAuthorizer;
 import com.softicar.platform.emf.authorizer.IEmfAuthorizer;
 import com.softicar.platform.emf.delete.IEmfDeleteStrategy;
-import com.softicar.platform.emf.form.EmfFormConfiguration;
-import com.softicar.platform.emf.form.configuration.IEmfFormFactory;
+import com.softicar.platform.emf.form.factory.IEmfFormFactory;
 import com.softicar.platform.emf.form.indicator.EmfFormIndicatorAlignment;
 import com.softicar.platform.emf.form.indicator.EmfFormIndicatorConfiguration;
 import com.softicar.platform.emf.form.indicator.IEmfFormIndicatorFactory;
@@ -205,7 +204,7 @@ public interface IEmfTable<R extends IEmfTableRow<R, P>, P, S> extends IDbTable<
 
 	default IEmfFormFactory<R> getFormFactory() {
 
-		return getEmfTableConfiguration().getFormConfiguration().getFormFactory();
+		return getEmfTableConfiguration().getFormFactory();
 	}
 
 	default Collection<IEmfFormIndicatorFactory<R>> getIndicatorFactories(EmfFormIndicatorAlignment alignment) {
@@ -357,11 +356,6 @@ public interface IEmfTable<R extends IEmfTableRow<R, P>, P, S> extends IDbTable<
 	default void customizeEmfTableConfiguration(EmfTableConfiguration<R, P, S> configuration) {
 
 		DevNull.swallow(configuration);
-	}
-
-	default void customizeForm(EmfFormConfiguration<R> formConfiguration) {
-
-		DevNull.swallow(formConfiguration);
 	}
 
 	default void customizeFormIndicators(EmfFormIndicatorConfiguration<R> indicatorConfiguration) {
