+++ platform-emf/src/main/java/com/softicar/platform/emf/table/configuration/IEmfTableConfiguration.java
@@ -23,7 +23,7 @@ import com.softicar.platform.emf.authorizer.IEmfAttributeAuthorizer;
 import com.softicar.platform.emf.authorizer.IEmfAuthorizer;
 import com.softicar.platform.emf.deactivation.IEmfTableRowDeactivationStrategy;
 import com.softicar.platform.emf.delete.IEmfDeleteStrategy;
-import com.softicar.platform.emf.form.configuration.IEmfFormConfiguration;
+import com.softicar.platform.emf.form.factory.IEmfFormFactory;
 import com.softicar.platform.emf.form.indicator.IEmfFormIndicatorConfiguration;
 import com.softicar.platform.emf.form.popup.IEmfFormPopupConfiguration;
 import com.softicar.platform.emf.form.section.IEmfFormSectionConfiguration;
@@ -50,8 +50,6 @@ public interface IEmfTableConfiguration<R extends IEmfTableRow<R, P>, P, S> {
 	IResource getIcon();
 
 	Function<R, IDomElement> getDisplayFactory();
-
-	IEmfFormConfiguration<R> getFormConfiguration();
 
 	IEmfFormIndicatorConfiguration<R> getFormIndicatorConfiguration();
 
@@ -119,6 +117,8 @@ public interface IEmfTableConfiguration<R extends IEmfTableRow<R, P>, P, S> {
 	Collection<IEmfValidator<R>> getValidators();
 
 	IEmfAuthorizer<R, S> getAuthorizer();
+
+	IEmfFormFactory<R> getFormFactory();
 
 	Function<S, DomPopup> getCreationPopupFactory();
